升级抽奖业务，英雄装备激活接口新增校验子级装备是否已激活功能，新增英雄升星接口。封装更新战力(判断该英雄是否在上阵中, 只有上阵中的英雄更新战力及矿工)公共接口。优化用户注册接口。完善英雄详情接口。升级优化定时器监听补单功能。封装公共获取装备合成项。升级优化战斗核心系统。

+++ gm-domain/src/main/java/com/gm/modules/order/entity/TransactionOrderEntity.java
@@ -57,11 +57,11 @@ public class TransactionOrderEntity implements Serializable {
 	/**
 	 * 消耗金额
 	 */
-	private Double transactionFee;
+	private BigDecimal transactionFee;
 	/**
 	 * GAS费
 	 */
-	private Long transactionGasFee;
+	private BigDecimal transactionGasFee;
 	/**
 	 * 抽奖类型('1':单抽，'2':十连抽)
 	 */
